generalized seq lag poly chkpt
